refactor : App 상태관리 Enum 추가
