+++ src/main/java/com/example/lv3/enums/AppStatus.java
@@ -1,0 +1,6 @@
+package com.example.lv3.enums;
+
+public enum AppStatus {
+    IN_PROGRESS,
+    FINISH
+}
